requirements.txt
Flask==2.2.5
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==0.21.0
gunicorn==22.0.0
flask-jwt-extended==4.6.0
pytest==7.4.0
pytest-cov==5.0.0
pylint==2.16.2